+++ src/System.Diagnostics.Process/tests/ProcessTests.Windows.cs
@@ -1,0 +1,19 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.IO;
+
+namespace System.Diagnostics.Tests
+{
+    public partial class ProcessTests
+    {
+        private string WriteScriptFile(string directory, string name, int returnValue)
+        {
+            string filename = Path.Combine(directory, name);
+            filename += ".bat";
+            File.WriteAllText(filename, $"exit {returnValue}");
+            return filename;
+        }
+    }
+}
